+++ Cell2FireC/Forest.cpp
@@ -17,8 +17,8 @@ using namespace std;
 /*
     Constructor
 */
-Forest::Forest(int _id, std::string _location, std::vector<int> _coord, int _nCells, float _area, double _vol,
-               double _age, float _perimeter, std::unordered_map<std::string, int> &_fTypes)
+Forest::Forest(int _id, std::string _location, std::vector<int> _coord, int _nCells, double _area, double _vol,
+               double _age, double _perimeter, std::unordered_map<std::string, int> &_fTypes)
 {
 
     // Populate fields
