+++ Cell2FireC/Forest.h
@@ -20,10 +20,10 @@ class Forest
 public:
     int id;
     int nCells;
-    float area;
+    double area;
     double vol;
     double age;
-    float perimeter;
+    double perimeter;
     std::string location;
     std::vector<int> coord;
     std::unordered_map<std::string, int> fTypes;
@@ -31,8 +31,8 @@ public:
     std::vector<int> burntCells;
 
     // Constructor & methods
-    Forest(int _id, std::string _location, std::vector<int> _coord, int _nCells, float _area,
-           double _vol, double _age, float _perimeter, std::unordered_map<std::string, int> &_fTypes);
+    Forest(int _id, std::string _location, std::vector<int> _coord, int _nCells, double _area,
+           double _vol, double _age, double _perimeter, std::unordered_map<std::string, int> &_fTypes);
 
     void print_info();
 };
